rename RouteDescriptionGenerator to RouteDescriptionPostprocessor
 - move it to separate header and cpp file
 - rewrite Qt RouteDescriptionBuilder to use postprocessor ^^^ and remove duplicate code
