+++ libosmscout-client-qt/src/osmscout/RouteDescriptionBuilder.cpp
@@ -32,11 +32,12 @@
 
 namespace osmscout {
 
-RouteDescriptionBuilder::RouteDescriptionBuilder()
+RouteDescriptionBuilder::RouteDescriptionBuilder::Callback::Callback(QList<RouteStep> &routeSteps):
+  routeSteps(routeSteps)
 {
 }
 
-RouteDescriptionBuilder::~RouteDescriptionBuilder()
+RouteDescriptionBuilder::RouteDescriptionBuilder::Callback::~Callback()
 {
 }
 
@@ -216,9 +217,9 @@ static QString CrossingWaysDescriptionToString(const osmscout::RouteDescription:
 }
 
 
-void RouteDescriptionBuilder::DumpStartDescription(QList<RouteStep> &routeSteps,
-                                                   const osmscout::RouteDescription::StartDescriptionRef& startDescription,
-                                                   const osmscout::RouteDescription::NameDescriptionRef& nameDescription)
+void RouteDescriptionBuilder::Callback::OnStart(const RouteDescription::StartDescriptionRef& startDescription,
+                                                const RouteDescription::TypeNameDescriptionRef& /*typeNameDescription*/,
+                                                const RouteDescription::NameDescriptionRef& nameDescription)
 {
   QString startDesc;
   QString driveAlongDesc;
@@ -255,8 +256,7 @@ void RouteDescriptionBuilder::DumpStartDescription(QList<RouteStep> &routeSteps,
   }
 }
 
-void RouteDescriptionBuilder::DumpTargetDescription(QList<RouteStep> &routeSteps,
-                                                    const osmscout::RouteDescription::TargetDescriptionRef& targetDescription)
+void RouteDescriptionBuilder::Callback::OnTargetReached(const osmscout::RouteDescription::TargetDescriptionRef& targetDescription)
 {
   RouteStep targetReached("target");
 
@@ -274,11 +274,11 @@ void RouteDescriptionBuilder::DumpTargetDescription(QList<RouteStep> &routeSteps
   routeSteps.push_back(targetReached);
 }
 
-void RouteDescriptionBuilder::DumpTurnDescription(QList<RouteStep> &routeSteps,
-                                                  const osmscout::RouteDescription::TurnDescriptionRef& /*turnDescription*/,
-                                                  const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription,
-                                                  const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
-                                                  const osmscout::RouteDescription::NameDescriptionRef& nameDescription)
+void RouteDescriptionBuilder::Callback::OnTurn(const osmscout::RouteDescription::TurnDescriptionRef& /*turnDescription*/,
+                                               const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription,
+                                               const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
+                                               const RouteDescription::TypeNameDescriptionRef& /*typeNameDescription*/,
+                                               const osmscout::RouteDescription::NameDescriptionRef& nameDescription)
 {
   RouteStep   turn(TurnCommandType(directionDescription));
   turn.shortDescription=ShortTurnCommand(directionDescription);
@@ -304,9 +304,8 @@ void RouteDescriptionBuilder::DumpTurnDescription(QList<RouteStep> &routeSteps,
   routeSteps.push_back(turn);
 }
 
-void RouteDescriptionBuilder::DumpRoundaboutEnterDescription(QList<RouteStep> &routeSteps,
-                                                             const osmscout::RouteDescription::RoundaboutEnterDescriptionRef& /*roundaboutEnterDescription*/,
-                                                             const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription)
+void RouteDescriptionBuilder::Callback::OnRoundaboutEnter(const osmscout::RouteDescription::RoundaboutEnterDescriptionRef& /*roundaboutEnterDescription*/,
+                                                          const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription)
 {
   RouteStep   enter("enter-roundabout");
   QString crossingWaysString;
@@ -325,9 +324,8 @@ void RouteDescriptionBuilder::DumpRoundaboutEnterDescription(QList<RouteStep> &r
   routeSteps.push_back(enter);
 }
 
-void RouteDescriptionBuilder::DumpRoundaboutLeaveDescription(QList<RouteStep> &routeSteps,
-                                                             const osmscout::RouteDescription::RoundaboutLeaveDescriptionRef& roundaboutLeaveDescription,
-                                                             const osmscout::RouteDescription::NameDescriptionRef& nameDescription)
+void RouteDescriptionBuilder::Callback::OnRoundaboutLeave(const osmscout::RouteDescription::RoundaboutLeaveDescriptionRef& roundaboutLeaveDescription,
+                                                          const osmscout::RouteDescription::NameDescriptionRef& nameDescription)
 {
   RouteStep leave("leave-roundabout");
 
@@ -347,9 +345,8 @@ void RouteDescriptionBuilder::DumpRoundaboutLeaveDescription(QList<RouteStep> &r
   routeSteps.push_back(leave);
 }
 
-void RouteDescriptionBuilder::DumpMotorwayEnterDescription(QList<RouteStep> &routeSteps,
-                                                           const osmscout::RouteDescription::MotorwayEnterDescriptionRef& motorwayEnterDescription,
-                                                           const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription)
+void RouteDescriptionBuilder::Callback::OnMotorwayEnter(const osmscout::RouteDescription::MotorwayEnterDescriptionRef& motorwayEnterDescription,
+                                                        const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription)
 {
   RouteStep   enter("enter-motorway");
 
@@ -384,8 +381,9 @@ void RouteDescriptionBuilder::DumpMotorwayEnterDescription(QList<RouteStep> &rou
   routeSteps.push_back(enter);
 }
 
-void RouteDescriptionBuilder::DumpMotorwayChangeDescription(QList<RouteStep> &routeSteps,
-                                                            const osmscout::RouteDescription::MotorwayChangeDescriptionRef& motorwayChangeDescription)
+void RouteDescriptionBuilder::Callback::OnMotorwayChange(const osmscout::RouteDescription::MotorwayChangeDescriptionRef& motorwayChangeDescription,
+                                                         const RouteDescription::MotorwayJunctionDescriptionRef& /*motorwayJunctionDescription*/,
+                                                         const RouteDescription::DestinationDescriptionRef& /*crossingDestinationDescription*/)
 {
   RouteStep change("change-motorway");
 
@@ -406,10 +404,10 @@ void RouteDescriptionBuilder::DumpMotorwayChangeDescription(QList<RouteStep> &ro
   routeSteps.push_back(change);
 }
 
-void RouteDescriptionBuilder::DumpMotorwayLeaveDescription(QList<RouteStep> &routeSteps,
-                                                           const osmscout::RouteDescription::MotorwayLeaveDescriptionRef& motorwayLeaveDescription,
-                                                           const osmscout::RouteDescription::DirectionDescriptionRef& /*directionDescription*/,
-                                                           const osmscout::RouteDescription::NameDescriptionRef& nameDescription)
+void RouteDescriptionBuilder::Callback::OnMotorwayLeave(const RouteDescription::MotorwayLeaveDescriptionRef& motorwayLeaveDescription,
+                                                        const RouteDescription::MotorwayJunctionDescriptionRef& /*motorwayJunctionDescription*/,
+                                                        const RouteDescription::DirectionDescriptionRef& /*directionDescription*/,
+                                                        const RouteDescription::NameDescriptionRef& nameDescription)
 {
   RouteStep leave("leave-motorway");
 
@@ -436,8 +434,7 @@ void RouteDescriptionBuilder::DumpMotorwayLeaveDescription(QList<RouteStep> &rou
   routeSteps.push_back(leave);
 }
 
-void RouteDescriptionBuilder::DumpNameChangedDescription(QList<RouteStep> &routeSteps,
-                                                         const osmscout::RouteDescription::NameChangedDescriptionRef& nameChangedDescription)
+void RouteDescriptionBuilder::Callback::OnPathNameChange(const osmscout::RouteDescription::NameChangedDescriptionRef& nameChangedDescription)
 {
   RouteStep changed("name-change");
 
@@ -455,188 +452,18 @@ void RouteDescriptionBuilder::DumpNameChangedDescription(QList<RouteStep> &route
   routeSteps.push_back(changed);
 }
 
-bool RouteDescriptionBuilder::GenerateRouteStep(const osmscout::RouteDescription::Node &node,
-                                                QList<RouteStep> &routeSteps,
-                                                size_t &roundaboutCrossingCounter)
-{
-  osmscout::RouteDescription::DescriptionRef                 desc;
-  osmscout::RouteDescription::NameDescriptionRef             nameDescription;
-  osmscout::RouteDescription::DirectionDescriptionRef        directionDescription;
-  osmscout::RouteDescription::NameChangedDescriptionRef      nameChangedDescription;
-  osmscout::RouteDescription::CrossingWaysDescriptionRef     crossingWaysDescription;
+RouteDescriptionBuilder::RouteDescriptionBuilder()
+{}
 
-  osmscout::RouteDescription::StartDescriptionRef            startDescription;
-  osmscout::RouteDescription::TargetDescriptionRef           targetDescription;
-  osmscout::RouteDescription::TurnDescriptionRef             turnDescription;
-  osmscout::RouteDescription::RoundaboutEnterDescriptionRef  roundaboutEnterDescription;
-  osmscout::RouteDescription::RoundaboutLeaveDescriptionRef  roundaboutLeaveDescription;
-  osmscout::RouteDescription::MotorwayEnterDescriptionRef    motorwayEnterDescription;
-  osmscout::RouteDescription::MotorwayChangeDescriptionRef   motorwayChangeDescription;
-  osmscout::RouteDescription::MotorwayLeaveDescriptionRef    motorwayLeaveDescription;
-
-  desc=node.GetDescription(osmscout::RouteDescription::WAY_NAME_DESC);
-  if (desc) {
-    nameDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::NameDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::DIRECTION_DESC);
-  if (desc) {
-    directionDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::DirectionDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::WAY_NAME_CHANGED_DESC);
-  if (desc) {
-    nameChangedDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::NameChangedDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::CROSSING_WAYS_DESC);
-  if (desc) {
-    crossingWaysDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::CrossingWaysDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::NODE_START_DESC);
-  if (desc) {
-    startDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::StartDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::NODE_TARGET_DESC);
-  if (desc) {
-    targetDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::TargetDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::TURN_DESC);
-  if (desc) {
-    turnDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::TurnDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::ROUNDABOUT_ENTER_DESC);
-  if (desc) {
-    roundaboutEnterDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::RoundaboutEnterDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::ROUNDABOUT_LEAVE_DESC);
-  if (desc) {
-    roundaboutLeaveDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::RoundaboutLeaveDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::MOTORWAY_ENTER_DESC);
-  if (desc) {
-    motorwayEnterDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayEnterDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::MOTORWAY_CHANGE_DESC);
-  if (desc) {
-    motorwayChangeDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayChangeDescription>(desc);
-  }
-
-  desc=node.GetDescription(osmscout::RouteDescription::MOTORWAY_LEAVE_DESC);
-  if (desc) {
-    motorwayLeaveDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayLeaveDescription>(desc);
-  }
-
-  if (crossingWaysDescription &&
-      roundaboutCrossingCounter>0 &&
-      crossingWaysDescription->GetExitCount()>1) {
-    roundaboutCrossingCounter+=crossingWaysDescription->GetExitCount()-1;
-  }
-
-  if (startDescription) {
-    DumpStartDescription(routeSteps,
-                         startDescription,
-                         nameDescription);
-  }
-  else if (targetDescription) {
-    DumpTargetDescription(routeSteps,
-                          targetDescription);
-  }
-  else if (turnDescription) {
-    DumpTurnDescription(routeSteps,
-                        turnDescription,
-                        crossingWaysDescription,
-                        directionDescription,
-                        nameDescription);
-  }
-  else if (roundaboutEnterDescription) {
-    DumpRoundaboutEnterDescription(routeSteps,
-                                   roundaboutEnterDescription,
-                                   crossingWaysDescription);
-
-    roundaboutCrossingCounter=1;
-  }
-  else if (roundaboutLeaveDescription) {
-    DumpRoundaboutLeaveDescription(routeSteps,
-                                   roundaboutLeaveDescription,
-                                   nameDescription);
-
-    roundaboutCrossingCounter=0;
-  }
-  else if (motorwayEnterDescription) {
-    DumpMotorwayEnterDescription(routeSteps,
-                                 motorwayEnterDescription,
-                                 crossingWaysDescription);
-  }
-  else if (motorwayChangeDescription) {
-    DumpMotorwayChangeDescription(routeSteps,
-                                  motorwayChangeDescription);
-  }
-  else if (motorwayLeaveDescription) {
-    DumpMotorwayLeaveDescription(routeSteps,
-                                 motorwayLeaveDescription,
-                                 directionDescription,
-                                 nameDescription);
-  }
-  else if (nameChangedDescription) {
-    DumpNameChangedDescription(routeSteps,
-                               nameChangedDescription);
-  }
-  else {
-    return false;
-  }
-  return true;
-}
+RouteDescriptionBuilder::~RouteDescriptionBuilder()
+{}
 
 void RouteDescriptionBuilder::GenerateRouteSteps(const osmscout::RouteDescription &routeDescription,
                                                  QList<RouteStep> &routeSteps)
 {
-  size_t roundaboutCrossingCounter=0;
-  int    lastStepIndex=-1;
-
-  std::list<osmscout::RouteDescription::Node>::const_iterator prevNode=routeDescription.Nodes().end();
-  for (std::list<osmscout::RouteDescription::Node>::const_iterator node=routeDescription.Nodes().begin();
-       node!=routeDescription.Nodes().end();
-       ++node) {
-
-    if (!GenerateRouteStep(*node, routeSteps, roundaboutCrossingCounter)){
-      continue;
-    }
-
-    int currentStepIndex;
-
-    if (lastStepIndex>=0) {
-        currentStepIndex=lastStepIndex+1;
-    }
-    else {
-        currentStepIndex=0;
-    }
-
-    if (currentStepIndex>=0) {
-      routeSteps[currentStepIndex].distance=node->GetDistance();
-      routeSteps[currentStepIndex].time=node->GetTime();
-
-      if (prevNode!=routeDescription.Nodes().end() &&
-        (node->GetDistance()-prevNode->GetDistance()).AsMeter()!=0.0) {
-        routeSteps[currentStepIndex].distanceDelta=(node->GetDistance()-prevNode->GetDistance());
-      }
-
-      if (prevNode!=routeDescription.Nodes().end() &&
-          node->GetTime()-prevNode->GetTime()!=Duration::zero()) {
-        routeSteps[currentStepIndex].timeDelta=node->GetTime()-prevNode->GetTime();
-      }
-    }
-
-    lastStepIndex=routeSteps.size()-1;
-
-    prevNode=node;
-  }
+  RouteDescriptionPostprocessor postprocessor;
+  Callback callback(routeSteps);
+  postprocessor.GenerateDescription(routeDescription, callback);
 }
+
 }
