+++ libosmscout/src/meson.build
@@ -33,6 +33,7 @@ osmscoutSrc = [
             'src/osmscout/util/WorkQueue.cpp',
             'src/osmscout/util/TagErrorReporter.cpp',
             'src/osmscout/routing/Route.cpp',
+            'src/osmscout/routing/RouteDescriptionPostprocessor.cpp'
             'src/osmscout/routing/RouteData.cpp',
             'src/osmscout/routing/RouteNode.cpp',
             'src/osmscout/routing/RouteNodeDataFile.cpp',
