+++ libosmscout/src/osmscout/routing/Route.cpp
@@ -23,7 +23,6 @@
 
 #include <osmscout/system/Assert.h>
 #include <osmscout/system/Math.h>
-#include "../../../../Demos/src/Navigation.h"
 
 namespace osmscout {
 
@@ -571,277 +570,5 @@ namespace osmscout {
                        targetNodeIndex);
   }
 
-  RouteDescriptionGenerator::Callback::~Callback()
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::BeforeRoute()
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnStart(const RouteDescription::StartDescriptionRef& /*startDescription*/,
-                                                    const RouteDescription::TypeNameDescriptionRef& /*typeNameDescription*/,
-                                                    const RouteDescription::NameDescriptionRef& /*nameDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnTargetReached(const RouteDescription::TargetDescriptionRef& /*targetDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnTurn(const RouteDescription::TurnDescriptionRef& /*turnDescription*/,
-                                                   const RouteDescription::CrossingWaysDescriptionRef& /*crossingWaysDescription*/,
-                                                   const RouteDescription::DirectionDescriptionRef& /*directionDescription*/,
-                                                   const RouteDescription::TypeNameDescriptionRef& /*typeNameDescription*/,
-                                                   const RouteDescription::NameDescriptionRef& /*nameDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnRoundaboutEnter(const RouteDescription::RoundaboutEnterDescriptionRef& /*roundaboutEnterDescription*/,
-                                                              const RouteDescription::CrossingWaysDescriptionRef& /*crossingWaysDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnRoundaboutLeave(const RouteDescription::RoundaboutLeaveDescriptionRef& /*roundaboutLeaveDescription*/,
-                                                              const RouteDescription::NameDescriptionRef& /*nameDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnMotorwayEnter(const RouteDescription::MotorwayEnterDescriptionRef& /*motorwayEnterDescription*/,
-                                                            const RouteDescription::CrossingWaysDescriptionRef& /*crossingWaysDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnMotorwayChange(const RouteDescription::MotorwayChangeDescriptionRef& /*motorwayChangeDescription*/,
-                                                             const RouteDescription::MotorwayJunctionDescriptionRef& /*motorwayJunctionDescription*/,
-                                                             const RouteDescription::DestinationDescriptionRef& /*crossingDestinationDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnMotorwayLeave(const RouteDescription::MotorwayLeaveDescriptionRef& /*motorwayLeaveDescription*/,
-                                                            const RouteDescription::MotorwayJunctionDescriptionRef& /*motorwayJunctionDescription*/,
-                                                            const RouteDescription::DirectionDescriptionRef& /*directionDescription*/,
-                                                            const RouteDescription::NameDescriptionRef& /*nameDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnPathNameChange(const RouteDescription::NameChangedDescriptionRef& /*nameChangedDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnMaxSpeed(const RouteDescription::MaxSpeedDescriptionRef& /*maxSpeedDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::OnPOIAtRoute(const RouteDescription::POIAtRouteDescriptionRef& /*poiAtRouteDescription*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::BeforeNode(const RouteDescription::Node& /*node*/)
-  {
-    // no code
-  }
-
-  void RouteDescriptionGenerator::Callback::AfterNode(const RouteDescription::Node& /*node*/)
-  {
-    // no code
-  }
-
-
-  /**
-   * Evaluate the already postprocessed RouteDescription and call the given callback for
-   * node segments where something important happens or changes.
-   *
-   * @param description
-   * @param callback
-   */
-  void RouteDescriptionGenerator::GenerateDescription(const RouteDescription& description,
-                                                      RouteDescriptionGenerator::Callback& callback)
-  {
-    callback.BeforeRoute();
-
-    auto prevNode=description.Nodes().cend();
-    for (auto node=description.Nodes().cbegin();
-         node!=description.Nodes().cend();
-         ++node) {
-      osmscout::RouteDescription::DescriptionRef                 desc;
-      osmscout::RouteDescription::NameDescriptionRef             nameDescription;
-      osmscout::RouteDescription::DirectionDescriptionRef        directionDescription;
-      osmscout::RouteDescription::NameChangedDescriptionRef      nameChangedDescription;
-      osmscout::RouteDescription::CrossingWaysDescriptionRef     crossingWaysDescription;
-
-      osmscout::RouteDescription::StartDescriptionRef            startDescription;
-      osmscout::RouteDescription::TargetDescriptionRef           targetDescription;
-      osmscout::RouteDescription::TurnDescriptionRef             turnDescription;
-      osmscout::RouteDescription::RoundaboutEnterDescriptionRef  roundaboutEnterDescription;
-      osmscout::RouteDescription::RoundaboutLeaveDescriptionRef  roundaboutLeaveDescription;
-      osmscout::RouteDescription::MotorwayEnterDescriptionRef    motorwayEnterDescription;
-      osmscout::RouteDescription::MotorwayChangeDescriptionRef   motorwayChangeDescription;
-      osmscout::RouteDescription::MotorwayLeaveDescriptionRef    motorwayLeaveDescription;
-      osmscout::RouteDescription::MotorwayJunctionDescriptionRef motorwayJunctionDescription;
-      osmscout::RouteDescription::DestinationDescriptionRef      crossingDestinationDescription;
-      osmscout::RouteDescription::MaxSpeedDescriptionRef         maxSpeedDescription;
-      osmscout::RouteDescription::TypeNameDescriptionRef         typeNameDescription;
-      osmscout::RouteDescription::POIAtRouteDescriptionRef       poiAtRouteDescription;
-
-      desc=node->GetDescription(osmscout::RouteDescription::WAY_NAME_DESC);
-      if (desc) {
-        nameDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::NameDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::DIRECTION_DESC);
-      if (desc) {
-        directionDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::DirectionDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::WAY_NAME_CHANGED_DESC);
-      if (desc) {
-        nameChangedDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::NameChangedDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::CROSSING_WAYS_DESC);
-      if (desc) {
-        crossingWaysDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::CrossingWaysDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::NODE_START_DESC);
-      if (desc) {
-        startDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::StartDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::NODE_TARGET_DESC);
-      if (desc) {
-        targetDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::TargetDescription>(desc);
-      }
-
-
-      desc=node->GetDescription(osmscout::RouteDescription::TURN_DESC);
-      if (desc) {
-        turnDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::TurnDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::ROUNDABOUT_ENTER_DESC);
-      if (desc) {
-        roundaboutEnterDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::RoundaboutEnterDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::ROUNDABOUT_LEAVE_DESC);
-      if (desc) {
-        roundaboutLeaveDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::RoundaboutLeaveDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::MOTORWAY_ENTER_DESC);
-      if (desc) {
-        motorwayEnterDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayEnterDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::MOTORWAY_CHANGE_DESC);
-      if (desc) {
-        motorwayChangeDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayChangeDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::MOTORWAY_LEAVE_DESC);
-      if (desc) {
-        motorwayLeaveDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayLeaveDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::MOTORWAY_JUNCTION_DESC);
-      if (desc) {
-        motorwayJunctionDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MotorwayJunctionDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::CROSSING_DESTINATION_DESC);
-      if (desc) {
-        crossingDestinationDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::DestinationDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::WAY_MAXSPEED_DESC);
-      if (desc) {
-        maxSpeedDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::MaxSpeedDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::WAY_TYPE_NAME_DESC);
-      if (desc) {
-        typeNameDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::TypeNameDescription>(desc);
-      }
-
-      desc=node->GetDescription(osmscout::RouteDescription::POI_AT_ROUTE_DESC);
-      if (desc) {
-        poiAtRouteDescription=std::dynamic_pointer_cast<osmscout::RouteDescription::POIAtRouteDescription>(desc);
-      }
-
-      callback.BeforeNode(*node);
-
-      if (startDescription) {
-        callback.OnStart(startDescription,
-                         typeNameDescription,
-                         nameDescription);
-      }
-
-      if (targetDescription) {
-        callback.OnTargetReached(targetDescription);
-      }
-
-      if (turnDescription) {
-        callback.OnTurn(turnDescription,
-                        crossingWaysDescription,
-                        directionDescription,
-                        typeNameDescription,
-                        nameDescription);
-      }
-      else if (roundaboutEnterDescription) {
-        callback.OnRoundaboutEnter(roundaboutEnterDescription,
-                                   crossingWaysDescription);
-      }
-      else if (roundaboutLeaveDescription) {
-        callback.OnRoundaboutLeave(roundaboutLeaveDescription,
-                                   nameDescription);
-      }
-
-      if (motorwayEnterDescription) {
-        callback.OnMotorwayEnter(motorwayEnterDescription,
-                                 crossingWaysDescription);
-      }
-      else if (motorwayChangeDescription) {
-        callback.OnMotorwayChange(motorwayChangeDescription,
-                                  motorwayJunctionDescription,
-                                  crossingDestinationDescription);
-      }
-      else if (motorwayLeaveDescription) {
-        callback.OnMotorwayLeave(motorwayLeaveDescription,
-                                 motorwayJunctionDescription,
-                                 directionDescription,
-                                 nameDescription);
-      }
-      else if (nameChangedDescription) {
-        callback.OnPathNameChange(nameChangedDescription);
-      }
-
-      if (maxSpeedDescription) {
-        callback.OnMaxSpeed(maxSpeedDescription);
-      }
-
-      if (poiAtRouteDescription) {
-        callback.OnPOIAtRoute(poiAtRouteDescription);
-      }
-
-      callback.AfterNode(*node);
-
-      prevNode=node;
-    }
-  }
 }
